+++ app/src/main/java/kr/ac/pusan/walkover/autotrackingcctv/ui/adapter/OnCameraCardClickedListener.java
@@ -3,5 +3,5 @@ package kr.ac.pusan.walkover.autotrackingcctv.ui.adapter;
 import android.view.View;
 
 public interface OnCameraCardClickedListener {
-    void onClicked(View view, int cameraId);
+    void onClicked(View view, long cameraId);
 }
